fix(soma): corrige a funcionalidade de soma

+++ src/soma.c
@@ -1,4 +1,4 @@
 #include<stdio.h>
-int soma(int num1, int num2){
+float soma(float num1, float num2){
     return num1 + num2;
 }
